Add job trigger logic to pjutil
